Fix Java Implementation of Aspect Ratio Test
Java's Implementation of Rect does not include getters for the member variables for its height and width. Instead, they are left accessible via direct member references. Therefore, instead of a getter, the implementation should use a direct usage of the values and cast it to a double to make sure the aspect ratio uses floating-point arithmetic.